[MDAPI-113] [C++][Tools] Tools should report invalid event type dxfcpp::RuntimeException
