+++ CMakeLists.txt
@@ -199,6 +199,7 @@ set(dxFeedGraalCxxApi_InternalUtilsDebug_Sources
 )
 
 set(dxFeedGraalCxxApi_Exceptions_Sources
+        src/exceptions/RuntimeException.cpp
         src/exceptions/GraalException.cpp
         src/exceptions/JavaException.cpp
 )
