+++ include/dxfeed_graal_cpp_api/api.hpp
@@ -52,6 +52,7 @@ DXFCXX_DISABLE_MSC_WARNINGS_PUSH(4251 4996)
 
 #include "isolated/Isolated.hpp"
 
+#include "exceptions/RuntimeException.hpp"
 #include "exceptions/JavaException.hpp"
 #include "exceptions/GraalException.hpp"
