+++ samples/cpp/DxFeedConnect/src/main.cpp
@@ -91,12 +91,8 @@ int main(int argc, char *argv[]) {
         }
 
         std::cin.get();
-    } catch (const JavaException &e) {
-        std::cerr << e.what() << '\n';
-        std::cerr << e.getStackTrace() << '\n';
-    } catch (const GraalException &e) {
-        std::cerr << e.what() << '\n';
-        std::cerr << e.getStackTrace() << '\n';
+    } catch (const RuntimeException &e) {
+        std::cerr << e << '\n';
     }
 
     return 0;
