+++ samples/cpp/PrintQuoteEvents/src/main.cpp
@@ -34,11 +34,7 @@ int main(int argc, char *argv[]) {
         subscription->addSymbols(symbol);
 
         std::cin.get();
-    } catch (const JavaException &e) {
-        std::cerr << e.what() << '\n';
-        std::cerr << e.getStackTrace() << '\n';
-    } catch (const GraalException &e) {
-        std::cerr << e.what() << '\n';
-        std::cerr << e.getStackTrace() << '\n';
+    } catch (const RuntimeException &e) {
+        std::cerr << e << '\n';
     }
 }
