+++ tools/Tools/src/Connect/ConnectTool.hpp
@@ -209,12 +209,8 @@ struct ConnectTool {
             sub->addSymbols(symbols);
             endpoint->connect(args.address);
             std::this_thread::sleep_for(std::chrono::days(365));
-        } catch (const JavaException &e) {
-            std::cerr << e.what() << '\n';
-            std::cerr << e.getStackTrace() << '\n';
-        } catch (const GraalException &e) {
-            std::cerr << e.what() << '\n';
-            std::cerr << e.getStackTrace() << '\n';
+        } catch (const RuntimeException &e) {
+            std::cerr << e << '\n';
         }
     }
 };
