+++ tools/Tools/src/Dump/DumpTool.hpp
@@ -184,12 +184,8 @@ struct DumpTool {
                 outputEndpoint.value()->awaitProcessed();
                 outputEndpoint.value()->closeAndAwaitTermination();
             }
-        } catch (const JavaException &e) {
-            std::cerr << e.what() << '\n';
-            std::cerr << e.getStackTrace() << '\n';
-        } catch (const GraalException &e) {
-            std::cerr << e.what() << '\n';
-            std::cerr << e.getStackTrace() << '\n';
+        } catch (const RuntimeException &e) {
+            std::cerr << e << '\n';
         }
     }
 };
