+++ tools/Tools/src/PerfTest/PerfTestTool.hpp
@@ -271,12 +271,8 @@ struct PerfTestTool {
             endpoint->closeAndAwaitTermination();
 
             std::cout << hash << std::endl;
-        } catch (const JavaException &e) {
-            std::cerr << e.what() << '\n';
-            std::cerr << e.getStackTrace() << '\n';
-        } catch (const GraalException &e) {
-            std::cerr << e.what() << '\n';
-            std::cerr << e.getStackTrace() << '\n';
+        } catch (const RuntimeException &e) {
+            std::cerr << e << '\n';
         }
     }
 };
